# openapi.yaml
openapi: 3.1.0
info:
  title: demo-repo API
  version: 0.1.0
paths:
  /users:
    get:
      tags:
      - users
      summary: List Users
      operationId: list_users_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                title: Response List Users Users Get
    post:
      tags:
      - users
      summary: Create User
      operationId: create_user_users_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /users/{user_id}:
    get:
      tags:
      - users
      summary: Get User
      operationId: get_user_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /posts:
    get:
      tags:
      - posts
      summary: List Posts
      operationId: list_posts_posts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
                title: Response List Posts Posts Get
    post:
      tags:
      - posts
      summary: Create Post
      operationId: create_post_posts_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /posts/{post_id}:
    get:
      tags:
      - posts
      summary: Get Post
      operationId: get_post_posts__post_id__get
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /:
    get:
      summary: Root
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Post:
      properties:
        title:
          type: string
          title: Title
        body:
          type: string
          title: Body
        author_id:
          type: integer
          title: Author Id
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - title
      - body
      - author_id
      - id
      - created_at
      title: Post
    PostCreate:
      properties:
        title:
          type: string
          title: Title
        body:
          type: string
          title: Body
        author_id:
          type: integer
          title: Author Id
      type: object
      required:
      - title
      - body
      - author_id
      title: PostCreate
    User:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - name
      - email
      - id
      - created_at
      title: User
    UserCreate:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
      type: object
      required:
      - name
      - email
      title: UserCreate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError